refactor(matrix): Share matrix factory
